+++ fe/fe-core/src/test/java/com/starrocks/connector/iceberg/IcebergAwsClientFactoryTest.java
@@ -14,6 +14,7 @@
 
 package com.starrocks.connector.iceberg;
 
+import com.starrocks.connector.share.credential.AwsCredentialUtil;
 import com.starrocks.connector.share.credential.CloudConfigurationConstants;
 import com.starrocks.connector.share.iceberg.IcebergAwsClientFactory;
 import org.apache.iceberg.aws.AwsClientProperties;
@@ -77,27 +78,27 @@ public class IcebergAwsClientFactoryTest {
     @Test
     public void testEnsureSchemeInEndpoint() {
         // test endpoint without scheme
-        URI uriWithoutScheme = IcebergAwsClientFactory.ensureSchemeInEndpoint("s3.aa-bbbbb-3.amazonaws.com.cn");
+        URI uriWithoutScheme = AwsCredentialUtil.ensureSchemeInEndpoint("s3.aa-bbbbb-3.amazonaws.com.cn");
         Assertions.assertEquals("https://s3.aa-bbbbb-3.amazonaws.com.cn", uriWithoutScheme.toString());
 
         // test endpoint with scheme HTTPS
-        URI uriWithHttps = IcebergAwsClientFactory.ensureSchemeInEndpoint("https://s3.aa-bbbbb-3.amazonaws.com.cn");
+        URI uriWithHttps = AwsCredentialUtil.ensureSchemeInEndpoint("https://s3.aa-bbbbb-3.amazonaws.com.cn");
         Assertions.assertEquals("https://s3.aa-bbbbb-3.amazonaws.com.cn", uriWithHttps.toString());
 
         // test endpoint with scheme HTTP
-        URI uriWithHttp = IcebergAwsClientFactory.ensureSchemeInEndpoint("http://s3.aa-bbbbb-3.amazonaws.com.cn");
+        URI uriWithHttp = AwsCredentialUtil.ensureSchemeInEndpoint("http://s3.aa-bbbbb-3.amazonaws.com.cn");
         Assertions.assertEquals("http://s3.aa-bbbbb-3.amazonaws.com.cn", uriWithHttp.toString());
 
         // test endpoint with port but without scheme (the bug case)
         // URI.create("localhost:4566") incorrectly parses "localhost" as scheme and "4566" as path
-        URI uriWithPort = IcebergAwsClientFactory.ensureSchemeInEndpoint("localhost:4566");
+        URI uriWithPort = AwsCredentialUtil.ensureSchemeInEndpoint("localhost:4566");
         Assertions.assertEquals("https://localhost:4566", uriWithPort.toString());
         Assertions.assertEquals("https", uriWithPort.getScheme());
         Assertions.assertEquals("localhost", uriWithPort.getHost());
         Assertions.assertEquals(4566, uriWithPort.getPort());
 
         // test endpoint with port and path but without scheme
-        URI uriWithPortAndPath = IcebergAwsClientFactory.ensureSchemeInEndpoint("example.com:8080/path");
+        URI uriWithPortAndPath = AwsCredentialUtil.ensureSchemeInEndpoint("example.com:8080/path");
         Assertions.assertEquals("https://example.com:8080/path", uriWithPortAndPath.toString());
         Assertions.assertEquals("https", uriWithPortAndPath.getScheme());
         Assertions.assertEquals("example.com", uriWithPortAndPath.getHost());
